Updated replaceSpaces()
I've changed replaceSpaces() to something a little less verbose. I like the look of this code better, it's way more readable, but for anything more than one space in the string it's going to make more sense to use the method with StringBuilder because it's mutable instead of needing to overwrite the string each time a change needs to be made.